fix: keep Wing fast path platform scoped
