+++ context_request_test.go
@@ -80,20 +80,6 @@ func TestCtxResetDefersTransportRequestContextLookup(t *testing.T) {
 	}
 }
 
-func TestCtxResetWingUsesLazyRequestContext(t *testing.T) {
-	app := New()
-	c := newCtx(app)
-	req := &wingRequest{ctx: context.WithValue(context.Background(), contextValueKey("wing-lazy"), "ok")}
-	resp := acquireResponse()
-	defer releaseResponse(resp)
-
-	c.resetWing(resp, req)
-
-	if got := c.Context().Value(contextValueKey("wing-lazy")); got != "ok" {
-		t.Fatalf("Context value = %v, want lazy Wing request context value", got)
-	}
-}
-
 func TestServeHTTPContextUsesHTTPRequestContext(t *testing.T) {
 	const key contextValueKey = "http-request-value"
